Preserve query string parameters for redirect URL
Prior to this any parameters after the & were treated as parameters
for the logout URL itself and not the redirect URL. Properly encoding
the URLs ensures that the entire URL is preserved.

The tests were also not working well due to the assignment of window.location.href.
This change fixes that and better supports different types of tests.